refactor(components): replace FormToggle with ThemeSwitcher
